backend/requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.0.7
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-redis==5.4.0
django-rq==2.10.2
djangorestframework==3.15.2
ffmpeg-python==0.2.0
future==1.0.0
imageio-ffmpeg==0.5.1
numpy==2.0.1
pillow==10.4.0
psutil==6.0.0
redis==5.0.8
rq==1.16.2
setuptools==72.1.0
sqlparse==0.5.1